+++ docs/changelog.md
@@ -73,7 +73,10 @@
 
 ### sys::mem
 - rename `define_arena!` to `arena!`
+
+### sys::mem::view
 - update `Slice`: add methods: `get`, `get_mut`:
+- new items: `MemReplicaError`, `MemReplicaSlice`.
 
 ### sys::os
 - remove macros: `os_print!`, `os_println!`, `os_eprint!`, `os_eprintln!`.